Move video files cleanup into a delayed task
