+++ src/b_upload/__init__.py
@@ -26,7 +26,6 @@ from wtforms import (
 
 from .. import app, db
 from ..core.et import cancel_job
-from ..core.s3 import S3Transfer
 from ..methods import get_thumbnail_cdn_url
 from ..models import (
     Video,
@@ -36,6 +35,7 @@ from ..models import (
     VideoFrameAnalysis,
     TranscoderStatus,
 )
+from ..tasks.shared import delete_video_files
 from ..tasks.thumbnails import process_thumbnails, process_avatar
 from ..tasks.transcoder import start_transcoder_job
 
@@ -416,7 +416,9 @@ def delete_unpublished_video(video: Video):
         s3_upload_thumbnail_key=video.file_mapper.s3_upload_thumbnail_key,
     )
     delete_video_sql(video.id)
-    delete_video_files(video.id, file_mapper_obj)
+    # cleanup actual files in 1 hour
+    # in case of delete being triggered in the middle of post-processing
+    delete_video_files.apply_async((video.id, file_mapper_obj), countdown=3600)
     return make_response('', 200)
 
 
@@ -426,8 +428,9 @@ def delete_video_sql(video_id):
 
     # cancel transcoding jobs
     for job in db.session.query(TranscoderJob).filter_by(video_id=video.id):
-        if job.status == TranscoderStatus.processing:
-            cancel_job(job.id)
+        if job.status in (TranscoderStatus.pending, TranscoderStatus.processing):
+            with suppress(Exception):
+                cancel_job(job.id)
 
     # delete orphaned (non-cascading entries)
     db.session.query(TranscoderJob).filter_by(video_id=video_id).delete()
@@ -435,47 +438,6 @@ def delete_video_sql(video_id):
 
     db.session.delete(video)
     db.session.commit()
-
-
-def delete_video_files(video_id: str, file_mapper_obj: dict):
-    # delete upload files
-    s3 = boto3.client(
-        's3',
-        region_name=app.config['S3_UPLOADS_REGION'],
-        aws_access_key_id=app.config['AWS_MANAGER_PUBLIC_KEY'],
-        aws_secret_access_key=app.config['AWS_MANAGER_PRIVATE_KEY'],
-    )
-    s3.delete_objects(
-        Bucket=file_mapper_obj['s3_upload_bucket'],
-        Delete={
-            'Objects': [
-                {
-                    'Key': file_mapper_obj['s3_upload_video_key'],
-                },
-                {
-                    'Key': file_mapper_obj['s3_upload_thumbnail_key'],
-                },
-            ],
-        },
-    )
-
-    # delete video files
-    s3t = S3Transfer(
-        region_name=app.config['S3_VIDEOS_REGION'],
-        bucket_name=app.config['S3_VIDEOS_BUCKET'],
-    )
-    keys_to_delete = [
-        *s3t.ls(f'snapshots/{video_id}'),
-        *s3t.ls(f'thumbnails/{video_id}'),
-        *s3t.ls(f'v1/{video_id}'),
-    ]
-    if keys_to_delete:
-        s3t.client.delete_objects(
-            Bucket=app.config['S3_VIDEOS_BUCKET'],
-            Delete={
-                'Objects': [{'Key': x['Key']} for x in keys_to_delete],
-            },
-        )
 
 
 def s3_make_public(bucket_name: str, key: str):
